Major revamp for 2.0 firmware.
This also makes the API dramatically simpler, because the concept of a
gateway bulb is gone in the new firmware.

You now:

- ask the local network for bulbs
- send packets directly to each of those bulbs

With this new API, you'd now say:

    bulbs = lazylights.find_bulbs(expected_bulbs=2)
    # Fade all bulbs up to warm white, 90% brightness, over 30 seconds.
    lazylights.set_state(bulbs, 0, 0, 0.9, 2000, fade=30000)

That's it.

The bulbs are much more responsive with the new firmware, and this
package is much less complex -- no more threads, locking, juggling TCP
and UDP connections, etc.